src: split up different uses of SHA-256
src/fsverity.rs defines Sha256HashValue as an implementation of
FsVerityHashValue but we also use it for situations where we handle
SHA-256 digests that aren't fs-verity hash values.  This happens a lot
in the OCI code (to refer to images and layers) and correspondingly in
the splitstream code (which stores those things).

Introduce a util::Sha256Digest type, returned by util::parse_sha256()
and use it consistently for all of the cases that have nothing to do
with fs-verity.

parse_sha256() is still used in some places in the code for values that
are intended as fs-verity digests.  We'll start fixing that up in the
following commits when we add some helpers directly on the types
themselves.

Signed-off-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>